Add install-dockup.sh script
Copied/modified from the perlbrew installer, added an install script
which simply pulls the latest dockup.sh from github and runs its
self-install (i.e. --install option). This exists to be able to install
directly via a piped command, for example:

  wget -O - http://rapi.io/install-dockup.sh | bash

Note, the above URL is not yet active, but I plan to make it so. Still,
this is just for illustrative puroses only